requirements.txt
aws-cdk-lib==2.150.0
constructs>=10.0.0,<11.0.0
cdk-ecr-deployment==3.0.96  # other versions of this library might not be stable
